requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
Django==3.1.2
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
djangorestframework==3.12.1
idna==2.10
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mistune==0.8.4
Pygments==2.7.1
pytz==2020.1
requests==2.24.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.25.11
